Added support for PAL-60.
This is a mode with NTSC line timings, but PAL encoding.

In theory the matching new kernel is required, but in practice the old
kernel mapped all unknown modes to NTSC and the LCDC configuration for
PAL-60 is identical to NTSC, so it works with the old kernel as well.